requirements.txt
argparser
requests
python-dotenv
scapy